refactor: 업로드할 이미지 설정하는 기능 커스텀 훅으로 분리 (#155)
